Moved the generic xxe checks to its own class
